Enable predictive back for Android 14+
