+++ app/src/main/kotlin/com/absinthe/libchecker/ui/album/ComparisonActivity.kt
@@ -16,7 +16,6 @@ import android.view.Gravity
 import android.view.MenuItem
 import android.view.ViewGroup
 import android.widget.FrameLayout
-import androidx.activity.addCallback
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
@@ -119,9 +118,6 @@ class ComparisonActivity : BaseActivity<ActivityComparisonBinding>() {
   }
 
   private fun registerCallbacks() {
-    onBackPressedDispatcher.addCallback(this, true) {
-      finish()
-    }
     chooseApkResultLauncher =
       registerForActivityResult(ActivityResultContracts.GetContent()) {
         if (isLeftPartChoosing) {
