+++ app/src/main/kotlin/com/absinthe/libchecker/ui/snapshot/AlbumActivity.kt
@@ -5,7 +5,6 @@ import android.os.Bundle
 import android.view.ContextThemeWrapper
 import android.view.MenuItem
 import android.view.ViewGroup
-import androidx.activity.addCallback
 import androidx.activity.viewModels
 import androidx.appcompat.app.AlertDialog
 import androidx.lifecycle.lifecycleScope
@@ -42,9 +41,6 @@ class AlbumActivity : BaseActivity<ActivityAlbumBinding>() {
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     initView()
-    onBackPressedDispatcher.addCallback(this, true) {
-      finish()
-    }
   }
 
   private fun initView() {
