+++ app/src/main/res/values-v34/values.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+  <bool name="appOnBackInvokedCallbackEnabled">true</bool>
+</resources>
