+++ app/src/main/res/values/values.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+  <!-- disable OnBackInvokedCallback for Android 13 and below -->
+  <!-- for it causes flickering even with the developer option turned off (i.e. common user xp) -->
+  <bool name="appOnBackInvokedCallbackEnabled">false</bool>
+</resources>
